Error Handling and custom Exceptions

+++ db/db_user.py
@@ -1,5 +1,6 @@
 # Third Party
 from sqlalchemy.orm.session import Session
+from fastapi import HTTPException,status
 
 # Project Library
 from schemas import UserBase
@@ -21,10 +22,15 @@ def get_all_users(db:Session):
     return db.query(DbUser).all()
 
 def get_user(db:Session,id:int):
-    return db.query(DbUser).filter(DbUser.id==id).first()
+    user = db.query(DbUser).filter(DbUser.id==id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f"User with id {id} not found")
+    return user
 
 def update_user(db:Session,id:int,request:UserBase):
     user = db.query(DbUser).filter(DbUser.id==id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f"User with id {id} not found") 
     user.update({
         DbUser.username : request.username,
         DbUser.email : request.email,
@@ -35,6 +41,8 @@ def update_user(db:Session,id:int,request:UserBase):
 
 def delete_user(db:Session,id:int):
     user = db.query(DbUser).filter(DbUser.id==id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f"User with id {id} not found")
     db.delete(user)
     db.commit()
     return "User Deleted"
